reserish/requirementsx.txt
# Cleaned-up requirements.txt
django
sentence-transformers
docx2txt
PyPDF2
numpy
scikit-learn
requests
groq
python-dotenv
youtube-search-python
psycopg2-binary
razorpay
django-cors-headers
selenium
channels
django-celery-beat
django-celery-results
djangorestframework-simplejwt
djangorestframework
pdfplumber
pandas
reportlab
PyMuPDF
networkx==3.4.2
httpx==0.23.0
gunicorn
whitenoise
django_redis
